Improvements to C solver, so solves all puzzles
First of all, the program now implements a "guess
and check" algorithm to solve puzzles that the previous
version couldn't.  The Python version still needs to be updated.

Also possibles for ALL spaces now default to anything, not just
blank spaces.  This doesn't change the result, but simplifies
control flow.