Fixes to C version
Fixes to C version to make sure that everything works fine
despite the change to recursive methods.

Actually performs full copies of the sudoku on the stack now, since
rounds of intensive solving might make changes even in a recursive
branch that might later be pruned.

solve_sudoku() also checks the state of the sudoku for validity so
that we correctly abort an invalid recursive branch.