Add rust set_value method and get_indices function
To the basic Suduko add blocks, cols, and rows to keep track of which
values are still available for each of the constraints on a sudoku.

One method to set the value for a given location and update the
constraints at the same time.

Add a function to get the row, column, and block for any given sudoku
square.